리스팅 수정 완료

+++ api/controller/ListingController.js
@@ -1,4 +1,5 @@
 import Listing from "../models/listing.js";
+import { errorHandler } from "../utils/error.js";
 
 export const createListing = async (req, res, next) => {
     try {
@@ -21,10 +22,40 @@ export const getMyListing = async (req, res, next) => {
 }
 
 export const deleteMyListing = async (req, res, next) => {
+    const listing = await Listing.findById(req.params.id);
+    if(!listing) return next(errorHandler(404, '해당 리스팅을 찾을 수 없습니다.'))
     try {
         console.log(req.params.id);
         await Listing.findByIdAndDelete(req.params.id);
         res.status(200).json('리스팅 삭제 완료!');
+    } catch (error) {
+        next(error);
+    }
+}
+
+export const updateMyListing = async (req, res, next) => {
+    const listing = await Listing.findById(req.params.id);
+    if(!listing) return next(errorHandler(404, '해당 리스팅을 찾을 수 없습니다.'))
+    try {
+        const updateListing = await Listing.findByIdAndUpdate(
+            req.params.id,
+            req.body,
+            {new:true},
+        );
+        res.status(200).json(updateListing);
+    } catch (error) {
+        next(error);
+    }
+}
+
+export const getListing = async (req, res, next) => {
+    try {
+        const listing = await Listing.findById(req.params.id);
+
+        if(!listing){
+            return next(errorHandler(404, '리스팅을 찾을 수 없습니다.'));
+        }
+        res.status(200).json(listing);
     } catch (error) {
         next(error);
     }
